refactored debug code to fit with main
